requirements.txt
pyrogram
tgcrypto
tortoise-orm[accel]
aiomysql